+++ logs/runlog.log
@@ -1,2 +1,3 @@
-2023-05-07 12:59:51 (main.py) INFO: User clicked X button
-2023-05-07 12:59:52 (main.py) INFO: Exited program
+2023-05-07 13:16:49 (main.py) DEBUG: {'Issue': {'ID': 281, 'Name': 'Food poisoning', 'Accuracy': 90, 'Icd': 'A05;A02;A03;A04', 'IcdName': 'Other bacterial foodborne intoxications, not elsewhere classified;Other salmonella infections;Shigellosis;Other bacterial intestinal infections', 'ProfName': 'Foodborne illness', 'Ranking': 1}, 'Specialisation': [{'ID': 15, 'Name': 'General practice', 'SpecialistID': 0}, {'ID': 19, 'Name': 'Internal medicine', 'SpecialistID': 0}]}
+2023-05-07 13:17:33 (main.py) INFO: User clicked X button
+2023-05-07 13:17:35 (main.py) INFO: Exited program
